Rename the smtpClientHandshakeLogic as per convention
