+++ package.json
@@ -21,14 +21,7 @@
   "bugs": {
     "url": "https://github.com/charmalt/final-project-smtp/issues"
   },
-  "dependencies": {
-    "cookie-parser": "~1.4.3",
-    "debug": "~2.6.9",
-    "express": "~4.16.0",
-    "http-errors": "~1.6.2",
-    "jade": "~1.11.0",
-    "morgan": "~1.9.0"
-  },
+  "dependencies": {},
   "homepage": "https://github.com/charmalt/final-project-smtp#readme",
   "devDependencies": {
     "eslint": "^5.0.1",
@@ -39,6 +32,5 @@
     "eslint-plugin-standard": "^3.1.0",
     "jest": "^23.2.0",
     "standard": "^11.0.1"
-  },
-  "dependencies": {}
+  }
 }
